refactor(fetcher): put teams in it's own file
