+++ src/data-fetcher/src/models/mod.rs
@@ -1,0 +1,6 @@
+mod teams;
+
+pub use teams::{
+    Teams,
+    TeamsType
+};
